Huge Refactoring
Removed Encryption
Added IL2 Message Handler
TODO - handle channels and improve Radio UI

+++ DCS-SR-Client/Audio/Models/ClientAudio.cs
@@ -12,7 +12,7 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Client
         public double Frequency { get; internal set; }
         public short Modulation { get; internal set; }
         public float Volume { get; internal set; }
-        public uint UnitId { get; set; }
+        public long UnitId { get; set; }
         public RadioReceivingState RadioReceivingState { get; set; }
         public ulong PacketNumber { get; set; }
         public string OriginalClientGuid { get; internal set; }
